# openapi.yaml
openapi: "3.0.3"
info:
  title: gr33n API
  version: "0.2.0"
  description: |
    Local-first automated agriculture platform API.
    All endpoints run on localhost:8080 by default.

    **Authentication:**
    - Public routes require no auth.
    - Pi routes require `X-API-Key` header.
    - Dashboard routes require `Authorization: Bearer <JWT>`.
    - In `AUTH_MODE=dev` (dev-tagged build only), auth is bypassed.
    - `AUTH_MODE=auth_test` enforces the same auth as production on a dev-tagged build only.

    **Operator docs:** Phase 13 index — `docs/phase-13-operator-documentation.md`. **Phase 14 index** — `docs/phase-14-operator-documentation.md`. Farm audit trail — `docs/audit-events-operator-playbook.md`. Insert Commons pipeline — `docs/insert-commons-pipeline-runbook.md`; receiver — `docs/insert-commons-receiver-playbook.md`. Commons catalog (gr33n_inserts) — `docs/commons-catalog-operator-playbook.md`. Receipt storage cutover — `docs/receipt-storage-cutover-runbook.md`. Push alerts (FCM) — `docs/notifications-operator-playbook.md`. Domain module stubs (DB only; no API yet) — `docs/domain-modules-operator-playbook.md`. Optional mobile wrapper — `docs/mobile-distribution.md`.

    **Phase 13 API areas:** `organizations` (multi-farm grouping, usage summary), `farms` (Insert Commons opt-in/sync/history, farm-scoped audit events), `costs` / attachments (finance mappings, exports, receipts).
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0
servers:
...
paths:
...
  /sensors/{id}:
    get:
      tags: [sensors]
      summary: Get sensor by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "200":
          description: Sensor object
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sensor" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [sensors]
      summary: Soft-delete a sensor
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/ID" }
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Sensor:
      type: object
      properties:
        id: { type: integer }
        farm_id: { type: integer }
        zone_id: { type: integer, nullable: true }
        device_id: { type: integer, nullable: true }
        name: { type: string, example: Air Temp Indoor }
        sensor_type: { type: string, example: temperature }
        unit_id: { type: integer }
        hardware_identifier: { type: string, nullable: true }
        value_min_expected: { type: number, nullable: true }
        value_max_expected: { type: number, nullable: true }
        alert_threshold_low: { type: number, nullable: true }
        alert_threshold_high: { type: number, nullable: true }
        reading_interval_seconds: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
...